Re-enable customParams.
customParams is actually useful for us — links coming into our pywb
server are guaranteed to uniquely identify a .warc/.cdx by guid, so we
want to pass that guid along to our CDX server. We could cram it into
the filter param, but that’s supposed to be for filtering CDX lines
based on their content, right?