Fix live rewrite test
The test fetches 'http://example.com' and had hardcoded assumptions
about the value of the 'Content-Length' header in the response and
the casing of the header.

The expected content length is for the uncompressed content size.
However, the current version of the requests lib sets the
'Accept-Encoding: gzip' header by default and also lowercases
the names of headers in the response.

The proper approach here would be to mock out the HTTP request.
This takes a slightly uglier initial step of performing a direct
request and comparing the X-Archive-Content-Length header
in the proxy's response to that of the original URL.