Enable Sauce Connect in Travis
Rather than specifying the username and encrypted
key in the Travis config, we just enable Sauce Connect
and require SAUCE_USERNAME and SAUCE_ACCESS_KEY env vars
to be set.

This is so that the Karma tests have the same env vars
available to them which they can use to check whether
to run against Sauce Labs or not.