Update test_html_rewriter to match Python HtmlParser's new behavior (#970)
In Python 3.9.24, 3.10.19, 3.11.14, 3.12.12, 3.13.6 and 3.14.0  HtmlParser was changed as a security update to more correctly follow the HTML5 standard. This changed the output of one of our tests. The output change doesn't appear to have a meaningful impact in this case as browsers treat it as a comment either way, so this just updates the test to match the output of the latest Python versions.

Fixes #969